Make constraint solving in bind_self() more principled
