+++ mypy/checker.py
@@ -48,7 +48,7 @@ from mypy.checkmember import (
 from mypy.typeops import (
     map_type_from_supertype, bind_self, erase_to_bound, make_simplified_union,
     erase_def_to_union_or_bound, erase_to_union_or_bound,
-    true_only, false_only, function_type,
+    true_only, false_only, function_type, TypeVarExtractor
 )
 from mypy import message_registry
 from mypy.subtypes import (
@@ -4525,20 +4525,6 @@ def detach_callable(typ: CallableType) -> CallableType:
         ret_type=type_list[-1],
     )
     return out
-
-
-class TypeVarExtractor(TypeQuery[List[TypeVarType]]):
-    def __init__(self) -> None:
-        super().__init__(self._merge)
-
-    def _merge(self, iter: Iterable[List[TypeVarType]]) -> List[TypeVarType]:
-        out = []
-        for item in iter:
-            out.extend(item)
-        return out
-
-    def visit_type_var(self, t: TypeVarType) -> List[TypeVarType]:
-        return [t]
 
 
 def overload_can_never_match(signature: CallableType, other: CallableType) -> bool:
